Add a symbol-purge endpoint; fix @Modifying return-type bug
- New DELETE /api/market-data/{symbol}: permanently drops a symbol's
  cached price_bar rows (distinct from /api/universe removal, which
  only stops it being traded/scanned but keeps the cache).
- Fix PriceBarRepository.deleteBySymbol: declared to return long, but
  Spring Data JPA's @Modifying bulk queries only accept void/int —
  compiles fine, fails at first invocation. Every deleteBySymbol call
  (refresh, ensureSymbol, and this new purge) was silently broken
  since the delete-ordering fix that introduced it. Matches the
  working int pattern already used by UniverseSymbolRepository's own
  deleteBySymbol. Verified against the live database this time, not
  just compiled.
- UniverseService.DEFAULTS: add RKLB, TSLA to the seeded default set.

+++ src/main/java/com/quant/finance/decision/repository/PriceBarRepository.java
@@ -25,7 +25,7 @@ public interface PriceBarRepository extends JpaRepository<PriceBarEntity, Long> 
      */
     @Modifying
     @Query("delete from PriceBarEntity b where b.symbol = :symbol")
-    long deleteBySymbol(@Param("symbol") String symbol);
+    int deleteBySymbol(@Param("symbol") String symbol);
 
     /** Drop every cached bar not fetched at the given timeframe (legacy nulls included). */
     @Modifying
